requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
certifi==2025.11.12
    # via requests
charset-normalizer==3.4.4
    # via requests
colorama==0.4.6
    # via tqdm
filelock==3.20.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2025.12.0
    # via
    #   huggingface-hub
    #   torch
huggingface-hub==0.36.0
    # via
    #   tokenizers
    #   transformers
idna==3.11
    # via requests
jinja2==3.1.6
    # via torch
markupsafe==3.0.3
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.6.1
    # via torch
numpy==2.4.0
    # via
    #   -r requirements.in
    #   torchvision
    #   transformers
packaging==25.0
    # via
    #   huggingface-hub
    #   transformers
pillow==12.0.0
    # via
    #   -r requirements.in
    #   torchvision
pyyaml==6.0.3
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   transformers
regex==2025.11.3
    # via transformers
requests==2.32.5
    # via
    #   huggingface-hub
    #   transformers
safetensors==0.7.0
    # via
    #   -r requirements.in
    #   transformers
sympy==1.14.0
    # via torch
tokenizers==0.22.1
    # via
    #   -r requirements.in
    #   transformers
torch==2.9.1
    # via
    #   -r requirements.in
    #   torchvision
torchvision==0.24.1
    # via -r requirements.in
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   transformers
transformers==4.57.3
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   huggingface-hub
    #   torch
urllib3==2.6.2
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
